Prepare v0.3.0 release: cleanup, module splits, and documentation
- Bump version to 0.3.0 with Keep a Changelog entry and rewritten release notes
- Split nova/evaluators/llm.py (1,372 lines) into the nova/evaluators/llm/
  package with per-provider modules; public import path unchanged
- Move NovaRuleFileParser to nova/core/rule_file.py (re-exported from parser)
- Split tests/test_sdk.py into engine and component test modules
- Add OpenRouter attribution headers (OPENROUTER_HTTP_REFERER/APP_TITLE)
- Fix CLI startup so keyword-only scans skip optional transformers/torch
- Remove legacy manual test harnesses, root test.py scratch script, and
  unused first.gif; update lint/compile gates across CI and docs
- Add ARCHITECTURE.md, INSTALLATION troubleshooting guide, README SDK
  quick-start and testing sections
- Detection verified byte-identical to previous code across 80 rule/prompt
  combinations; 151 tests, lint, build, twine, artifact, and wheel smoke
  gates all pass
