+++ nova/_version.py
@@ -1,3 +1,3 @@
 """Package version metadata."""
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
